soundness: Bruteforce all basic Barcan instances
Since we now know that Barcan and its converse are not valid (and so
neither are the two other Barcan-like formulas that may be obtained from
their contrapositives), the question is:

Is there any Barcan-like formula that *is* valid?

A Barcan-like formula is one that commutes one of the binders (∀, ∃) with
an operator or its dual (σ, σᵈ), in any of the operator's argument (any
any φᵢ with 1 ≤ i ≤ n).

This commit introduces a simple Python script that generates
soundness & unsoundness statements for all basic Barcan-like formulas,
imports them into Lean, and does the proofs for Barcan11Unsound and
Barcan12Unsound.